+++ Source/EasyNetQ/ComponentRegistration.cs
@@ -3,6 +3,7 @@ using EasyNetQ.AutoSubscribe;
 using EasyNetQ.Consumer;
 using EasyNetQ.Loggers;
 using EasyNetQ.Producer;
+using EasyNetQ.Rpc;
 
 namespace EasyNetQ
 {
