--- Source/EasyNetQ/Rpc/IRpcHeaderKeys.cs
@@ -1,8 +1,0 @@
-﻿namespace EasyNetQ.Rpc
-{
-    interface IRpcHeaderKeys
-    {
-        string IsFaultedKey { get; }
-        string ExceptionMessageKey { get; }
-    }
-}
